changed default stream to <<stdout+'\n'>>
changed to code so that the default stream is stdout with a line break at the end. 
Before this change the default stream was stdout and the line break was _hard coded_ during the write operation. 

This prevented using a custom streams which do not require or expect line breaks (e.g., winston transports)