Updating code example for logging on rotation
'rotating-file-stream' api has changed and most people will install the latest. So the example should adapt to that.
Also the previous code example was incomplete and left many people wondering about details of implementation.
My example is complete and short. Anyone can copy-paste it and get it immediately working. No need to go to the 'rotating-file-stream' library and read its documentation!